Refactor: 관리자용 메서드 분리
