+++ src/main/java/com/jandi/plan_backend/user/service/UserService.java
@@ -243,60 +243,4 @@ public class UserService {
     public boolean isExistUserName(String userName) {
         return userRepository.findByUserName(userName).isPresent();
     }
-
-    /** 관리자 전용 */
-    //유저 목록 로드
-    public Page<UserListDTO> getAllUsers(String userEmail, int page, int size) {
-        User admin = validationUtil.validateUserExists(userEmail);
-        validationUtil.validateUserIsAdmin(admin);
-
-        long totalCount = userRepository.count();
-        return PaginationService.getPagedData(page, size, totalCount,
-                userRepository::findAll, UserListDTO::new);
-    }
-
-    //부적절 유저 목록 로드
-    public Page<UserListDTO> getRestrictedUsers(String userEmail, int page, int size) {
-        User admin = validationUtil.validateUserExists(userEmail);
-        validationUtil.validateUserIsAdmin(admin);
-
-        long totalCount = userRepository.countByReportedIsTrue();
-        return PaginationService.getPagedData(page, size, totalCount,
-                userRepository::findByReportedIsTrue, UserListDTO::new);
-    }
-
-    //유저 제재하기
-    public Boolean permitUser(String userEmail, Integer userId) {
-        User admin = validationUtil.validateUserExists(userEmail);
-        validationUtil.validateUserIsAdmin(admin);
-
-        //유저 찾기
-        User user = userRepository.findByUserId(userId).orElse(null);
-        if(user==null) { //사용자 미존재 시 오류 반환
-            throw new BadRequestExceptionMessage("사용자가 존재하지 않습니다");
-        }
-        log.info("선택된 유저: {}{}", user.getEmail(), user.getReported() ? "(제재)" : "(일반)");
-
-        //유저 제재/제한 해제 후 제한 여부 반환
-        user.setReported(!user.getReported());
-        userRepository.save(user);
-        return user.getReported();
-    }
-
-    //아직 게시물 탈퇴 등은 이루어지지 않음!
-    public Boolean withdrawUser(String userEmail, Integer userId) {
-        User admin = validationUtil.validateUserExists(userEmail);
-        validationUtil.validateUserIsAdmin(admin);
-
-        //유저 찾기
-        User user = userRepository.findByUserId(userId).orElse(null);
-        if(user==null) { //사용자 미존재 시 오류 반환
-            throw new BadRequestExceptionMessage("사용자가 존재하지 않습니다");
-        }
-        log.info("선택된 유저: {}{}", user.getEmail(), user.getReported() ? "(제재)" : "(일반)");
-
-        //탈퇴 후 유저 존재여부 반환(탈퇴가 잘 되었다면 true)
-        deleteUser(user.getEmail());
-        return userRepository.findByUserId(userId).isEmpty();
-    }
 }
